+++ packages/react-hook/src/createUseStorageState.ts
@@ -1,0 +1,84 @@
+// Token from @alibaba/hooks (MIT)
+import { isFunction, isUndefined } from '@minko-fe/lodash-pro'
+import { useMemoizedFn, useUpdateEffect } from 'ahooks'
+import { useState } from 'react'
+
+export type SetState<S> = S | ((prevState?: S) => S)
+
+export interface Options<T> {
+  defaultValue?: T | (() => T)
+  serializer?: (value: T) => string
+  deserializer?: (value: string) => T
+  onError?: (error: unknown) => void
+}
+
+export function createUseStorageState(getStorage: () => Storage | undefined) {
+  function useStorageState<T>(key: string, options: Options<T> = {}) {
+    let storage: Storage | undefined
+    const {
+      onError = (e) => {
+        console.error(e)
+      },
+    } = options
+
+    // https://github.com/alibaba/hooks/issues/800
+    try {
+      storage = getStorage()
+    } catch (err) {
+      onError(err)
+    }
+
+    const serializer = (value: T) => {
+      if (options.serializer) {
+        return options.serializer(value)
+      }
+      return JSON.stringify(value)
+    }
+
+    const deserializer = (value: string): T => {
+      if (options.deserializer) {
+        return options.deserializer(value)
+      }
+      return JSON.parse(value)
+    }
+
+    function getStoredValue() {
+      try {
+        const raw = storage?.getItem(key)
+        if (raw) {
+          return deserializer(raw)
+        }
+      } catch (e) {
+        onError(e)
+      }
+      if (isFunction(options.defaultValue)) {
+        return options.defaultValue()
+      }
+      return options.defaultValue
+    }
+
+    const [state, setState] = useState(getStoredValue)
+
+    useUpdateEffect(() => {
+      setState(getStoredValue())
+    }, [key])
+
+    const updateState = (value?: SetState<T>) => {
+      const currentState = isFunction(value) ? value(state) : value
+      setState(currentState)
+
+      if (isUndefined(currentState)) {
+        storage?.removeItem(key)
+      } else {
+        try {
+          storage?.setItem(key, serializer(currentState))
+        } catch (e) {
+          console.error(e)
+        }
+      }
+    }
+
+    return [state, useMemoizedFn(updateState)] as const
+  }
+  return useStorageState
+}