+++ packages/react-hook/src/useLocalStorageSetState.ts
@@ -1,0 +1,16 @@
+import { type Options } from './createUseStorageState'
+import { type RequireSome } from './types'
+import { type StateType, useSetState } from './useSetState'
+
+export function useLocalStorageSetState<T extends StateType>(
+  key: string,
+  options: RequireSome<Options<T>, 'defaultValue'>,
+) {
+  return useSetState(options.defaultValue!, {
+    storage: {
+      api: localStorage,
+      key,
+      options: options as Options<T>, // Explicitly cast options to Options<T | undefined>
+    },
+  })
+}