+++ playground/spa/vite.config.ts
@@ -1,18 +1,19 @@
 import { i18nDetector } from '@minko-fe/react-locale/plugin'
 import { injectScripts, overrideConfig } from '@minko-fe/vite-config'
-import { Env } from '@minko-fe/vite-config/client'
 import react from '@vitejs/plugin-react'
 import { defineConfig } from 'vite'
 
 // https://vitejs.dev/config/
 export default defineConfig(async (env) => {
-  console.log(Env.development)
   return overrideConfig(
     env,
     {
       plugins: [
         react(),
-        i18nDetector(),
+        i18nDetector({
+          root: __dirname,
+          autoDetectI18nConfig: true,
+        }),
         injectScripts((manifest) => [
           {
             attrs: {