+++ packages/react-hook/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @minko-fe/react-hook
 
+## 0.2.0
+
+### Minor Changes
+
+- add `useSetState` plugin
+  refactor deps
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @minko-fe/lodash-pro@0.2.0
+
 ## 0.1.17
 
 ### Patch Changes